fix(video): close production e2e gaps

+++ scripts/render_ktd_six_names_hybrid.py
@@ -13,7 +13,6 @@ import argparse
 import hashlib
 import json
 import math
-import shutil
 import subprocess
 import tempfile
 from dataclasses import dataclass
@@ -175,7 +174,7 @@ def render_frames(shots: list[Shot], temp_dir: Path, fps: int, size: tuple[int, 
     frame_index = 0
     for shot in shots:
         source = cache.setdefault(shot.image, Image.open(shot.image))
-        frame_count = max(1, int(round((shot.end - shot.start) * fps)))
+        frame_count = max(1, round((shot.end - shot.start) * fps))
         for local_index in range(frame_count):
             progress = local_index / max(1, frame_count - 1)
             frame = animated_frame(source, progress, shot.motion, shot.direction, size)
